bugfix: 大盘变量all和reg的逻辑修正
