+++ src/pages/dashboard/VariableConfig/DisplayItem.tsx
@@ -4,7 +4,7 @@ import { resourceGroupItem } from '@/store/businessInterface';
 import { useTranslation } from 'react-i18next';
 import { FormType } from './EditItem';
 import { Variable } from './definition';
-import { convertExpressionToQuery, replaceExpressionVars } from './constant';
+import { convertExpressionToQuery, replaceExpressionVars, stringToRegex } from './constant';
 const { Option } = Select;
 import { Range } from '@/components/DateRangePicker';
 interface Props {
@@ -14,43 +14,6 @@ interface Props {
   data: Variable[];
   range: Range;
   onChange: (index: number, value: string | string[], options?) => void;
-}
-
-// const stringToRegex = (str) => {
-//   // Main regex
-//   const main = str.match(/\/(.+)\/.*/)[1];
-
-//   // Regex options
-//   const options = str.match(/\/.+\/(.*)/)[1];
-
-//   // Compiled regex
-//   return new RegExp(main, options);
-// };
-
-export function stringStartsAsRegEx(str: string): boolean {
-  if (!str) {
-    return false;
-  }
-
-  return str[0] === '/';
-}
-
-export function stringToRegex(str: string): RegExp | false {
-  if (!stringStartsAsRegEx(str)) {
-    return new RegExp(`^${str}$`);
-  }
-
-  const match = str.match(new RegExp('^/(.*?)/(g?i?m?y?)$'));
-
-  // if (!match) {
-  //   throw new Error(`'${str}' is not a valid regular expression.`);
-  // }
-
-  if (match) {
-    return new RegExp(match[1], match[2]);
-  } else {
-    return false;
-  }
 }
 
 const DisplayItem: React.FC<Props> = ({ expression, index, data, onChange, cluster, range }) => {
@@ -68,13 +31,14 @@ const DisplayItem: React.FC<Props> = ({ expression, index, data, onChange, clust
         setRange(range);
         setCurCluster(cluster);
         convertExpressionToQuery(newExpression, range).then((res) => {
-          setOptions(res);
           // 逻辑上只有导入大盘后初始化那一次 selected会为空
+          const regFilterRes = res.filter((i) => !reg || !stringToRegex(reg) || (stringToRegex(reg) as RegExp).test(i));
+          setOptions(regFilterRes);
           if (res.length > 0 && !selected) {
-            onChange(index, multi ? [res[0]] : res[0], res);
+            onChange(index, multi ? [regFilterRes[0]] : regFilterRes[0], regFilterRes);
           }
           if (exp && newExpression && exp !== newExpression) {
-            onChange(index, multi ? [] : '', res);
+            onChange(index, multi ? [] : '', regFilterRes);
           }
         });
       }
@@ -119,7 +83,7 @@ const DisplayItem: React.FC<Props> = ({ expression, index, data, onChange, clust
             )}
             {options &&
               options
-                .filter((i) => !reg || !stringToRegex(reg) || (stringToRegex(reg) as RegExp).test(i))
+                // .filter((i) => !reg || !stringToRegex(reg) || (stringToRegex(reg) as RegExp).test(i))
                 .map((value) => (
                   <Option key={value} value={value}>
                     {value}
@@ -130,7 +94,7 @@ const DisplayItem: React.FC<Props> = ({ expression, index, data, onChange, clust
           <AutoComplete style={{ width: 180 }} onChange={(v) => onChange(index, v)} placeholder='input here' value={selected as string} dropdownClassName='overflow-586'>
             {options &&
               options
-                .filter((i) => !reg || !stringToRegex(reg) || (stringToRegex(reg) as RegExp).test(i))
+                // .filter((i) => !reg || !stringToRegex(reg) || (stringToRegex(reg) as RegExp).test(i))
                 .map((value) => (
                   <Option key={value} value={value}>
                     {value}
